+++ src/mo_config/flags.ml
@@ -63,6 +63,7 @@ let force_gc = ref false
 let global_timer = ref true
 let experimental_field_aliasing = ref false
 let ocaml_js = ref false
+let js_project_root : string option ref = ref None
 let rts_stack_pages_default = 32 (* 2MB *)
 let rts_stack_pages : int option ref = ref None
 let rtti = ref false